db-api/requirements.txt
Flask==2.3.2
fastapi==0.115.0
gunicorn==21.2.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pydantic==2.8.2